backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
anthropic==0.86.0
openai==1.58.1
pillow==12.1.1
python-multipart==0.0.22
sqlalchemy==2.0.48
pydantic==2.12.5
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
mercadopago==2.3.0
email-validator==2.2.0
opencv-python-headless==4.13.0.92
numpy>=2.0.0
apscheduler==3.11.2
supabase==2.10.0
psycopg2-binary==2.9.10
